i hate it

+++ app/src/main/java/com/theonlylies/musictagger/utils/BitmapCache.java
@@ -3,7 +3,6 @@ package com.theonlylies.musictagger.utils;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.net.Uri;
-import android.util.Log;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -12,7 +11,6 @@ import java.io.IOException;
 import java.nio.channels.FileChannel;
 
 import io.reactivex.Observable;
-import io.reactivex.schedulers.Schedulers;
 
 public class BitmapCache {
     public BitmapCache(Context context) {
@@ -37,7 +35,7 @@ public class BitmapCache {
 
     public Uri cacheBitmap(Bitmap bitmap) throws Exception {
 
-        File cachedFile = new File(rootDir.getAbsolutePath() + File.pathSeparator + System.currentTimeMillis());
+        cachedFile = new File(rootDir.getAbsolutePath() + File.pathSeparator + System.currentTimeMillis());
         if (cachedFile.createNewFile()) {
             bitmap.compress(Bitmap.CompressFormat.PNG, 100, new FileOutputStream(cachedFile));
         }
@@ -47,19 +45,19 @@ public class BitmapCache {
 
     public Observable<Uri> reactiveCacheBitmap(Bitmap bitmap) throws Exception {
 
-        File cachedFile = new File(rootDir.getAbsolutePath() + File.pathSeparator + System.currentTimeMillis());
+        cachedFile = new File(rootDir.getAbsolutePath() + File.pathSeparator + System.currentTimeMillis());
         return Observable.fromCallable(() -> {
-            Log.e("rxjava", "observe");
             if (cachedFile.createNewFile()) {
                 bitmap.compress(Bitmap.CompressFormat.PNG, 100, new FileOutputStream(cachedFile));
                 return Uri.fromFile(cachedFile);
             } else return null;
-        }).subscribeOn(Schedulers.newThread());
+        });
 
     }
 
 
     public void clearCache() {
-        cachedFile.delete();
+        if (cachedFile != null)
+            cachedFile.delete();
     }
 }
